Added mutate_rows function from Davis Vaughan and updated the examineChoices function to circumvent the following issue:
There is a reinforcing image, 7006, that is in the IAPS Key but the new cohort subjects were not asked to rate it in the new ratings task for whatever reason. It was, however, rated by the original cohort. In the examineChoices function, we run into an issue with the subjects who were reinforced with this image and comparing them with the ones who weren’t. Until we address this, we are imputing the positive and negative values for each subject on this image by using the mean positive and negative ratings for it from the original cohort.